move RunXxxExampleIT tests from example to examples
