+++ contribs/drt/src/test/java/org/matsim/contrib/drt/run/examples/RunDrtExampleIT.java
@@ -20,7 +20,7 @@
 /**
  *
  */
-package org.matsim.contrib.drt.run.example;
+package org.matsim.contrib.drt.run.examples;
 
 import java.util.Collections;
 import java.util.Set;
@@ -31,7 +31,6 @@ import org.matsim.contrib.drt.data.DrtRequest;
 import org.matsim.contrib.drt.data.validator.DrtRequestValidator;
 import org.matsim.contrib.drt.run.DrtConfigGroup;
 import org.matsim.contrib.drt.run.DrtControlerCreator;
-import org.matsim.contrib.drt.run.examples.RunDrtExample;
 import org.matsim.contrib.dvrp.run.DvrpConfigGroup;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
